1.2.5

+++ Back_end/cadastro/src/main/java/com/Infinity/cadastro/repository/pessoaRepo.java
@@ -3,6 +3,6 @@ package com.Infinity.cadastro.repository;
 import com.Infinity.cadastro.model.pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface pessoaRepo extends JpaRepository<pessoa, Long> {
+public interface pessoaRepo extends JpaRepository<pessoa, Integer> {
 
 }
